More torch fixes
